Create the tensor from shm  with torch.frombuffer.
